DROOLS-1080 Test to reproduce ProjectClassLoader.getResources("") with kie-ci does not return all valid urls
